requirements.txt
keras
opencv-python
Pillow
adafruit-io
python-dotenv
pyserial